#ROBOMERGE-OWNER: benn.gallagher
#ROBOMERGE-AUTHOR: benn.gallagher
#ROBOMERGE-SOURCE: CL 16268359 in //UE5/Main/... via CL 16272209
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Test -> Release-Engine-Staging) (v799-16237190)

[CL 16272225 by benn gallagher in ue5-release-engine-staging branch]

+++ Engine/Source/Runtime/Experimental/Chaos/Private/GeometryCollection/GeometryCollectionClusteringUtility.cpp
@@ -130,7 +130,7 @@ void FGeometryCollectionClusteringUtility::ClusterAllBonesUnderNewRoot(FGeometry
 		}
 		ExplodedTransforms[RootNoneIndex] = Transforms[RootNoneIndex];
 		// This bones offset is the average of all the selected bones
-		ExplodedVectors[RootNoneIndex] = SumOfOffsets / ChildBones.Num();
+		ExplodedVectors[RootNoneIndex] = SumOfOffsets / static_cast<float>(ChildBones.Num());
 	}
 
 	// Selected Bone Setup
